requirement.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
